% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_all.R
\name{calc_all}
\alias{calc_all}
\title{Calculate Ratio, Ln(Ratio), and Ratio(-Ln(Ratio)) values}
\usage{
calc_all(famgenspe = famgenspe)
}
\arguments{
\item{famgenspe}{The example data table.}
}
\value{
Ratio(-Ln(Ratio)) values of every genus of family.
}
\description{
Formula: Ratio = (N_species_in_genus) / (N_species_in_family), Ln(Ratio) = Ln((N_species_in_genus) / (N_species_in_family)), NegMulRatio = Ratio * (-Ln((N_species_in_genus) / (N_species_in_family)))
}
\examples{
data("famgenspe")
all_ratio <- calc_all(famgenspe = famgenspe)

}
\author{
benben-miao
}
