% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ln_ratio.R
\name{calc_ln_ratio}
\alias{calc_ln_ratio}
\title{Calculate Ln(Ratio) values}
\usage{
calc_ln_ratio(famgenspe = famgenspe)
}
\arguments{
\item{famgenspe}{The example data table.}
}
\value{
Ln(Ratio) values of every genus of family.
}
\description{
Formula: Ln(Ratio) = Ln((N_species_in_genus) / (N_species_in_family))
}
\examples{
data("famgenspe")
ln_ratio <- calc_ln_ratio(famgenspe = famgenspe)

}
\author{
benben-miao
}
