% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mul_ratio.R
\name{calc_mul_ratio}
\alias{calc_mul_ratio}
\title{Calculate Ratio(-Ln(Ratio)) values}
\usage{
calc_mul_ratio(famgenspe = famgenspe)
}
\arguments{
\item{famgenspe}{The example data table.}
}
\value{
Ratio(-Ln(Ratio)) values of every genus of family.
}
\description{
Formula: Ratio = Ratio * (-Ln((N_species_in_genus) / (N_species_in_family)))
}
\examples{
data("famgenspe")
mul_ratio <- calc_mul_ratio(famgenspe = famgenspe)

}
\author{
benben-miao
}
