% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ratio.R
\name{calc_ratio}
\alias{calc_ratio}
\title{Calculate Ratio values}
\usage{
calc_ratio(famgenspe = famgenspe)
}
\arguments{
\item{famgenspe}{The example data table.}
}
\value{
Ratio values of every genus of family.
}
\description{
Formula: Ratio = (N_species_in_genus) / (N_species_in_family)
}
\examples{
data("famgenspe")
ratio <- calc_ratio(famgenspe = famgenspe)

}
\author{
benben-miao
}
