% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{famgenspe}
\alias{famgenspe}
\title{Example data for FishRatioR}
\format{
Columns: family (string), genus (string), species (numeric).
}
\usage{
data(famgenspe)
}
\description{
This example data include three columns: family (string), genus (string), species (numeric).
}
\examples{
data(famgenspe)

}
\references{
Satpathy A T, Granja J M, Yost K E, et al. (2019) Nature biotechnology 37, 925–936.
(\href{https://www.ncbi.nlm.nih.gov/pubmed/31375813}{PubMed})
}
\keyword{datasets}
